Refactor Link
