+++ src/entity/mod.rs
@@ -2,6 +2,7 @@ mod active_model;
 mod base_entity;
 mod column;
 mod identity;
+mod link;
 mod model;
 pub mod prelude;
 mod primary_key;
@@ -11,6 +12,7 @@ pub use active_model::*;
 pub use base_entity::*;
 pub use column::*;
 pub use identity::*;
+pub use link::*;
 pub use model::*;
 // pub use prelude::*;
 pub use primary_key::*;
