+++ src/entity/relation.rs
@@ -30,22 +30,6 @@ where
     }
 }
 
-pub trait Linked {
-    type FromEntity: EntityTrait;
-
-    type ToEntity: EntityTrait;
-
-    fn link(&self) -> Vec<RelationDef>;
-
-    fn find_linked(&self) -> Select<Self::ToEntity> {
-        let mut select = Select::new();
-        for rel in self.link().into_iter().rev() {
-            select = select.join_rev(JoinType::InnerJoin, rel);
-        }
-        select
-    }
-}
-
 #[derive(Debug)]
 pub struct RelationDef {
     pub rel_type: RelationType,
